+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ## main
 <!--- - <description> (#<PR_number, kudos to @<author>) --->
+- Change Reqres class name to ReqresLogger ([#36](https://github.com/AckeeCZ/Reqres/pull/36), kudos to @olejnjak)
 
 ## 3.2.0